+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-12-19 11:29:14
+
+### Changed Files
+- `supabase/migrations/20251219_001_fix_contract_level_and_earned_propagation.sql`
+
+### Statistics
+```
+ 1 file changed, 261 insertions(+)
+```
+
 ## [Checkpoint] - 2025-12-19 09:36:18
 
 ### Changed Files